requirements.txt
streamlit
supabase
openai
google-api-python-client
google-auth
